sessions/04_embedding_data_visualization_and_cluster_check/requirements.txt
google-cloud-bigquery[bqstorage,pandas]
google-auth
polars
pyarrow
pandas-gbq
scikit-learn
umap-learn
hdbscan
matplotlib